+++ ContentPatcher/Framework/ContentPatcherAPI.cs
@@ -4,7 +4,6 @@ using System.Text.RegularExpressions;
 using ContentPatcher.Framework.Tokens;
 using ContentPatcher.Framework.Tokens.ValueProviders;
 using ContentPatcher.Framework.Tokens.ValueProviders.ModConvention;
-using Pathoschild.Stardew.Common.Utilities;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework
@@ -30,9 +29,6 @@ namespace ContentPatcher.Framework
         /// <summary>A regex pattern matching a valid token name.</summary>
         private readonly Regex ValidNameFormat = new Regex("^[a-z]+$", RegexOptions.IgnoreCase);
 
-        /// <summary>A cache of logged deprecation warnings.</summary>
-        private readonly InvariantHashSet LoggedDeprecationWarnings = new InvariantHashSet();
-
 
         /*********
         ** Public methods
@@ -57,32 +53,6 @@ namespace ContentPatcher.Framework
         public void RegisterToken(IManifest mod, string name, Func<IEnumerable<string>> getValue)
         {
             this.RegisterValueProvider(mod, new ModSimpleValueProvider(name, getValue));
-        }
-
-        /// <summary>Register a complex token.</summary>
-        /// <param name="mod">The manifest of the mod defining the token (see <see cref="Mod.ModManifest"/> on your entry class).</param>
-        /// <param name="name">The token name. This only needs to be unique for your mod; Content Patcher will prefix it with your mod ID automatically, like <c>Pathoschild.ExampleMod/SomeTokenName</c>.</param>
-        /// <param name="updateContext">A function which updates the token value (if needed), and returns whether the token changed. Content Patcher will call this method once when it's updating the context (e.g. when a new day starts). The token is 'changed' if it may return a different value *for the same inputs* than before; it's important to report a change correctly, since Content Patcher will use this to decide whether patches need to be rechecked.</param>
-        /// <param name="isReady">A function which returns whether the token is available for use. This is always called after <paramref name="updateContext"/>. If this returns false, any patches or dynamic tokens using this token will be disabled. (A token may return true and still have no value, in which case the token value is simply blank.)</param>
-        /// <param name="getValue">A function which returns the current value for the given input arguments (if any). For example, <c>{{your-mod-id/PlayerInitials}}</c> would result in a null input; <c>{{your-mod-id/PlayerInitials:{{PlayerName}}}}</c> would pass in the parsed string after token substitution, like <c>"John Smith"</c>. If the token doesn't use input arguments, you can simply ignore the input.</param>
-        /// <param name="allowsInput">Whether the player can provide input arguments (see <paramref name="getValue"/>).</param>
-        /// <param name="requiresInput">Whether the token can *only* be used with input arguments (see <paramref name="getValue"/>).</param>
-        [Obsolete]
-        public void RegisterToken(IManifest mod, string name, Func<bool> updateContext, Func<bool> isReady, Func<string, IEnumerable<string>> getValue, bool allowsInput, bool requiresInput)
-        {
-            // log deprecation warning
-            string warning = $"{mod.Name} used an experimental token API that will break in the next Content Patcher update.";
-            if (this.LoggedDeprecationWarnings.Add(warning))
-                this.Monitor.Log(warning, LogLevel.Warn);
-
-            // hack to wrap legacy token
-            ConventionWrapper wrapper = new ConventionWrapper();
-            this.Reflection.GetField<ConventionDelegates.UpdateContext>(wrapper, $"{nameof(wrapper.UpdateContext)}Impl").SetValue(() => updateContext());
-            this.Reflection.GetField<ConventionDelegates.IsReady>(wrapper, $"{nameof(wrapper.IsReady)}Impl").SetValue(() => isReady());
-            this.Reflection.GetField<ConventionDelegates.GetValues>(wrapper, $"{nameof(wrapper.GetValues)}Impl").SetValue(input => getValue(input));
-            this.Reflection.GetField<ConventionDelegates.AllowsInput>(wrapper, $"{nameof(wrapper.AllowsInput)}Impl").SetValue(() => allowsInput);
-            this.Reflection.GetField<ConventionDelegates.RequiresInput>(wrapper, $"{nameof(wrapper.RequiresInput)}Impl").SetValue(() => requiresInput);
-            this.RegisterValueProvider(mod, new ConventionValueProvider(name, wrapper));
         }
 
         /// <summary>Register a complex token. This is an advanced API; only use this method if you've read the documentation and are aware of the consequences.</summary>