+++ ContentPatcher/Framework/ModTokenContext.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using ContentPatcher.Framework.Conditions;
 using ContentPatcher.Framework.Tokens;
+using ContentPatcher.Framework.Tokens.ValueProviders;
 using Pathoschild.Stardew.Common.Utilities;
 
 namespace ContentPatcher.Framework
@@ -18,23 +20,25 @@ namespace ContentPatcher.Framework
         /// <summary>The parent token context.</summary>
         private readonly IContext ParentContext;
 
-        /// <summary>The available local standard tokens.</summary>
+        /// <summary>The standard local tokens for this content pack.</summary>
+        /// <remarks>This includes config tokens and per-mod core tokens like <see cref="ConditionType.HasFile"/>.</remarks>
         private readonly GenericTokenContext LocalContext;
 
-        /// <summary>The dynamic tokens whose value depends on <see cref="DynamicTokenValues"/>.</summary>
-        private readonly GenericTokenContext<DynamicToken> DynamicContext;
+        /// <summary>The dynamic tokens for this content pack.</summary>
+        private readonly GenericTokenContext DynamicContext;
 
-        /// <summary>The conditional values used to set the values of <see cref="DynamicContext"/> tokens.</summary>
+        /// <summary>The dynamic tokens stored in the <see cref="DynamicContext"/>.</summary>
+        private readonly InvariantDictionary<ManagedManualToken> DynamicTokens = new InvariantDictionary<ManagedManualToken>();
+
+        /// <summary>The possible values for the <see cref="DynamicTokens"/>.</summary>
+        /// <remarks>These must be stored in registration order, since each token value may affect the value of subsequent tokens.</remarks>
         private readonly IList<DynamicTokenValue> DynamicTokenValues = new List<DynamicTokenValue>();
-
-        /// <summary>The underlying token contexts in priority order.</summary>
-        private readonly IContext[] Contexts;
 
         /// <summary>Maps tokens to those affected by changes to their value in the mod context.</summary>
         private InvariantDictionary<InvariantHashSet> TokenDependents { get; } = new InvariantDictionary<InvariantHashSet>();
 
-        /// <summary>The new tokens which haven't received a context update yet.</summary>
-        private readonly InvariantHashSet PendingTokens = new InvariantHashSet();
+        /// <summary>Whether any tokens haven't received a context update yet.</summary>
+        private bool HasNewTokens;
 
 
         /*********
@@ -51,13 +55,12 @@ namespace ContentPatcher.Framework
             this.Scope = scope;
             this.ParentContext = parentContext;
             this.LocalContext = new GenericTokenContext(this.IsModInstalled);
-            this.DynamicContext = new GenericTokenContext<DynamicToken>(this.IsModInstalled);
-            this.Contexts = new[] { this.ParentContext, this.LocalContext, this.DynamicContext };
+            this.DynamicContext = new GenericTokenContext(this.IsModInstalled);
         }
 
         /// <summary>Add a standard token to the context.</summary>
         /// <param name="token">The config token to add.</param>
-        public void Add(IHigherLevelToken<IToken> token)
+        public void AddLocalToken(IToken token)
         {
             if (token.Scope != this.Scope)
                 throw new InvalidOperationException($"Can't register the '{token.Name}' mod token because its scope '{token.Scope}' doesn't match this mod scope '{this.Scope}.");
@@ -69,98 +72,108 @@ namespace ContentPatcher.Framework
                 throw new InvalidOperationException($"The '{token.Name}' token is already registered.");
 
             this.LocalContext.Save(token);
-            this.PendingTokens.Add(token.Name);
+            this.HasNewTokens = true;
         }
 
         /// <summary>Add a dynamic token value to the context.</summary>
-        /// <param name="tokenValue">The token to add.</param>
-        public void Add(DynamicTokenValue tokenValue)
+        /// <param name="name">The token name.</param>
+        /// <param name="rawValue">The token value to set.</param>
+        /// <param name="conditions">The conditions that must match to set this value.</param>
+        public void AddDynamicToken(string name, IManagedTokenString rawValue, IEnumerable<Condition> conditions)
         {
             // validate
-            if (this.ParentContext.Contains(tokenValue.Name, enforceContext: false))
-                throw new InvalidOperationException($"Can't register a '{tokenValue}' token because there's a global token with that name.");
-            if (this.LocalContext.Contains(tokenValue.Name, enforceContext: false))
-                throw new InvalidOperationException($"Can't register a '{tokenValue.Name}' dynamic token because there's a config token with that name.");
+            if (this.ParentContext.Contains(name, enforceContext: false))
+                throw new InvalidOperationException($"Can't register a '{name}' token because there's a global token with that name.");
+            if (this.LocalContext.Contains(name, enforceContext: false))
+                throw new InvalidOperationException($"Can't register a '{name}' dynamic token because there's a config token with that name.");
 
             // get (or create) token
-            DynamicToken token;
+            if (!this.DynamicTokens.TryGetValue(name, out ManagedManualToken managed))
             {
-                if (!this.DynamicContext.Tokens.TryGetValue(tokenValue.Name, out IHigherLevelToken<DynamicToken> wrapper))
-                    this.DynamicContext.Save(wrapper = new HigherLevelTokenWrapper<DynamicToken>(new DynamicToken(tokenValue.Name, this.Scope)));
-                token = wrapper.Token;
+                managed = new ManagedManualToken(name, this.Scope);
+                this.DynamicTokens[name] = managed;
+                this.DynamicContext.Save(managed.Token);
             }
 
-            // add token value
-            token.AddTokensUsed(tokenValue.GetTokensUsed());
-            token.AddAllowedValues(tokenValue.Value);
+            // create token value handler
+            var tokenValue = new DynamicTokenValue(managed, rawValue, conditions);
+            string[] tokensUsed = tokenValue.GetTokensUsed().ToArray();
+
+            // save value info
+            managed.ValueProvider.AddTokensUsed(tokensUsed);
+            managed.ValueProvider.AddAllowedValues(rawValue);
             this.DynamicTokenValues.Add(tokenValue);
 
             // track tokens which should trigger an update to this token
-            Queue<string> tokenQueue = new Queue<string>(tokenValue.GetTokensUsed());
+            Queue<string> tokenQueue = new Queue<string>(tokensUsed);
             InvariantHashSet visited = new InvariantHashSet();
             while (tokenQueue.Any())
             {
                 // get token name
-                string tokenName = tokenQueue.Dequeue();
-                if (!visited.Add(tokenName))
+                string usedTokenName = tokenQueue.Dequeue();
+                if (!visited.Add(usedTokenName))
                     continue;
 
-                // if the current token uses other tokens, they may affect the being added too
-                IToken curToken = this.GetToken(tokenName, enforceContext: false);
-                foreach (string name in curToken.GetTokensUsed())
-                    tokenQueue.Enqueue(name);
-                if (curToken is DynamicToken curDynamicToken)
-                {
-                    foreach (string name in curDynamicToken.GetPossibleTokensUsed())
-                        tokenQueue.Enqueue(name);
-                }
+                // if the used token uses other tokens, they may affect the one being added too
+                IToken usedToken = this.GetToken(usedTokenName, enforceContext: false);
+                foreach (string nextTokenName in usedToken.GetTokensUsed())
+                    tokenQueue.Enqueue(nextTokenName);
 
-                // add dynamic value as a dependency of the current token
-                if (!this.TokenDependents.TryGetValue(curToken.Name, out InvariantHashSet used))
-                    this.TokenDependents.Add(curToken.Name, used = new InvariantHashSet());
-                used.Add(tokenValue.Name);
+                // add new token as a dependent of the used token
+                if (!this.TokenDependents.TryGetValue(usedToken.Name, out InvariantHashSet used))
+                    this.TokenDependents.Add(usedToken.Name, used = new InvariantHashSet());
+                used.Add(name);
             }
+
+            // track new token
+            this.HasNewTokens = true;
         }
 
         /// <summary>Update the current context.</summary>
         /// <param name="globalChangedTokens">The global token values which changed.</param>
         public void UpdateContext(InvariantHashSet globalChangedTokens)
         {
-            // nothing to do
-            if (!globalChangedTokens.Any() && !this.PendingTokens.Any())
-                return;
-
-            // get dynamic tokens which need an update
-            InvariantHashSet affectedTokens = this.GetTokensToUpdate(globalChangedTokens);
-            if (!affectedTokens.Any())
-                return;
-
             // update local standard tokens
-            foreach (var token in this.LocalContext.Tokens.Values)
+            //
+            // Some local tokens may change independently (e.g. Random), so we need to update all
+            // standard tokens here.
+            bool localTokensChanged = false;
+            foreach (IToken token in this.LocalContext.GetTokens(enforceContext: false))
             {
-                if (token.IsMutable && affectedTokens.Contains(token.Name))
-                    token.UpdateContext(this);
+                if (token.IsMutable)
+                    localTokensChanged |= token.UpdateContext(this);
             }
 
             // reset dynamic tokens
-            // note: since token values are affected by the order they're defined, only updating tokens affected by globalChangedTokens is not trivial.
-            foreach (DynamicToken token in this.DynamicContext.Tokens.Values.Select(p => p.Token))
+            //
+            // Since dynamic token values are affected by the order they're defined (e.g. one
+            // dynamic token can use the value of another), only updating tokens affected by
+            // globalChangedTokens isn't trivial. Instead we track which global tokens are used
+            // indirectly through dynamic tokens via AddDynamicToken, and use that to decide which
+            // patches to update.
+            if (globalChangedTokens.Any() || localTokensChanged || this.HasNewTokens)
             {
-                token.SetValue(null);
-                token.SetReady(false);
-            }
-            foreach (DynamicTokenValue tokenValue in this.DynamicTokenValues)
-            {
-                tokenValue.UpdateContext(this);
-                if (tokenValue.IsReady && tokenValue.Conditions.All(p => p.IsMatch))
+                foreach (ManagedManualToken managed in this.DynamicTokens.Values)
                 {
-                    DynamicToken token = this.DynamicContext.Tokens[tokenValue.Name].Token;
-                    token.SetValue(tokenValue.Value);
-                    token.SetReady(true);
+                    managed.ValueProvider.SetValue(null);
+                    managed.ValueProvider.SetReady(false);
+                }
+
+                foreach (DynamicTokenValue tokenValue in this.DynamicTokenValues)
+                {
+                    tokenValue.UpdateContext(this);
+                    if (tokenValue.IsReady && tokenValue.Conditions.All(p => p.IsMatch))
+                    {
+                        ManualValueProvider valueProvider = tokenValue.ParentToken.ValueProvider;
+
+                        valueProvider.SetValue(tokenValue.Value);
+                        valueProvider.SetReady(true);
+                    }
                 }
             }
 
-            this.PendingTokens.Clear();
+            // reset tracking
+            this.HasNewTokens = false;
         }
 
         /// <summary>Get the tokens affected by changes to a given token.</summary>
@@ -184,13 +197,13 @@ namespace ContentPatcher.Framework
         /// <inheritdoc />
         public bool Contains(string name, bool enforceContext)
         {
-            return this.Contexts.Any(p => p.Contains(name, enforceContext));
+            return this.GetContexts().Any(p => p.Contains(name, enforceContext));
         }
 
         /// <inheritdoc />
         public IToken GetToken(string name, bool enforceContext)
         {
-            foreach (IContext context in this.Contexts)
+            foreach (IContext context in this.GetContexts())
             {
                 IToken token = context.GetToken(name, enforceContext);
                 if (token != null)
@@ -203,7 +216,7 @@ namespace ContentPatcher.Framework
         /// <inheritdoc />
         public IEnumerable<IToken> GetTokens(bool enforceContext)
         {
-            foreach (IContext context in this.Contexts)
+            foreach (IContext context in this.GetContexts())
             {
                 foreach (IToken token in context.GetTokens(enforceContext))
                     yield return token;
@@ -221,23 +234,12 @@ namespace ContentPatcher.Framework
         /*********
         ** Private methods
         *********/
-        /// <summary>Get the tokens which need a context update.</summary>
-        /// <param name="globalChangedTokens">The global token values which changed.</param>
-        private InvariantHashSet GetTokensToUpdate(InvariantHashSet globalChangedTokens)
+        /// <summary>Get the underlying contexts in priority order.</summary>
+        private IEnumerable<IContext> GetContexts()
         {
-            // add tokens which depend on a changed global token
-            InvariantHashSet affectedTokens = new InvariantHashSet();
-            foreach (string globalToken in globalChangedTokens)
-            {
-                foreach (string affectedToken in this.GetTokensAffectedBy(globalToken))
-                    affectedTokens.Add(affectedToken);
-            }
-
-            // add uninitialized tokens
-            foreach (string patch in this.PendingTokens)
-                affectedTokens.Add(patch);
-
-            return affectedTokens;
+            yield return this.ParentContext;
+            yield return this.LocalContext;
+            yield return this.DynamicContext;
         }
     }
 }