+++ ContentPatcher/Framework/Patches/LoadPatch.cs
@@ -13,21 +13,23 @@ namespace ContentPatcher.Framework.Patches
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="path">The path to the patch from the root content file.</param>
-        /// <param name="parentPatch">The parent patch for which this patch was loaded, if any.</param>
-        /// <param name="contentPack">The content pack which requested the patch.</param>
         /// <param name="assetName">The normalized asset name to intercept.</param>
-        /// <param name="conditions">The conditions which determine whether this patch should be applied.</param>
         /// <param name="localAsset">The asset key to load from the content pack instead.</param>
+        /// <param name="conditions">The conditions which determine whether this patch should be applied.</param>
+        /// <param name="updateRate">When the patch should be updated.</param>
+        /// <param name="contentPack">The content pack which requested the patch.</param>
+        /// <param name="parentPatch">The parent patch for which this patch was loaded, if any.</param>
         /// <param name="normalizeAssetName">Normalize an asset name.</param>
-        public LoadPatch(LogPathBuilder path, IPatch parentPatch, ManagedContentPack contentPack, IManagedTokenString assetName, IEnumerable<Condition> conditions, IManagedTokenString localAsset, Func<string, string> normalizeAssetName)
+        public LoadPatch(LogPathBuilder path, IManagedTokenString assetName, IManagedTokenString localAsset, IEnumerable<Condition> conditions, UpdateRate updateRate, ManagedContentPack contentPack, IPatch parentPatch, Func<string, string> normalizeAssetName)
             : base(
                   path: path,
                   type: PatchType.Load,
-                  contentPack: contentPack,
                   assetName: assetName,
                   conditions: conditions,
-                  normalizeAssetName: normalizeAssetName,
+                  updateRate: updateRate,
+                  contentPack: contentPack,
                   parentPatch: parentPatch,
+                  normalizeAssetName: normalizeAssetName,
                   fromAsset: localAsset
                 )
         { }