+++ ContentPatcher/ModEntry.cs
@@ -9,6 +9,7 @@ using ContentPatcher.Framework.Conditions;
 using ContentPatcher.Framework.ConfigModels;
 using ContentPatcher.Framework.Migrations;
 using ContentPatcher.Framework.Tokens;
+using ContentPatcher.Framework.Tokens.ValueProviders;
 using ContentPatcher.Framework.Validators;
 using Pathoschild.Stardew.Common.Utilities;
 using StardewModdingAPI;
@@ -48,7 +49,8 @@ namespace ContentPatcher
             new Migration_1_14(),
             new Migration_1_15_Prevalidation(),
             new Migration_1_15_Rewrites(content),
-            new Migration_1_16()
+            new Migration_1_16(),
+            new Migration_1_17()
         };
 
         /// <summary>The special validation logic to apply to assets affected by patches.</summary>
@@ -157,7 +159,7 @@ namespace ContentPatcher
                 case LoadStage.Loaded when Game1.dayOfMonth == 0: // handled by OnDayStarted if we're not creating a new save
                     this.Monitor.VerboseLog($"Updating context: load stage changed to {e.NewStage}.");
                     this.TokenManager.IsBasicInfoLoaded = true;
-                    this.UpdateContext();
+                    this.UpdateContext(ContextUpdateType.All);
                     break;
             }
         }
@@ -169,7 +171,7 @@ namespace ContentPatcher
         {
             this.Monitor.VerboseLog("Updating context: new day started.");
             this.TokenManager.IsBasicInfoLoaded = true;
-            this.UpdateContext();
+            this.UpdateContext(ContextUpdateType.All);
         }
 
         /// <summary>The method invoked when the player warps.</summary>
@@ -177,9 +179,8 @@ namespace ContentPatcher
         /// <param name="e">The event data.</param>
         private void OnWarped(object sender, WarpedEventArgs e)
         {
-            ConditionType[] affectedTokens = new[] { ConditionType.LocationName, ConditionType.IsOutdoors };
-            this.Monitor.VerboseLog($"Updating context for {string.Join(", ", affectedTokens)}: player warped.");
-            this.UpdateContext(affectedTokens);
+            this.Monitor.VerboseLog("Updating context: player warped.");
+            this.UpdateContext(ContextUpdateType.OnLocationChange);
         }
 
         /// <summary>The method invoked when the player returns to the title screen.</summary>
@@ -189,7 +190,7 @@ namespace ContentPatcher
         {
             this.Monitor.VerboseLog("Updating context: returned to title.");
             this.TokenManager.IsBasicInfoLoaded = false;
-            this.UpdateContext();
+            this.UpdateContext(ContextUpdateType.All);
         }
 
         /// <summary>Raised after the game performs its overall update tick (≈60 times per second).</summary>
@@ -236,12 +237,12 @@ namespace ContentPatcher
             // load content packs and context
             this.TokenManager = new TokenManager(helper.Content, installedMods, this.QueuedModTokens, this.Helper.Reflection);
             this.PatchManager = new PatchManager(this.Monitor, this.TokenManager, this.AssetValidators());
-            this.PatchLoader = new PatchLoader(this.PatchManager, this.TokenManager, this.Monitor, installedMods, this.Helper.Content.NormalizeAssetName);
-            this.UpdateContext(); // set initial context before loading any custom mod tokens
+            this.PatchLoader = new PatchLoader(this.PatchManager, this.TokenManager, this.Monitor, this.Helper.Reflection, installedMods, this.Helper.Content.NormalizeAssetName);
+            this.UpdateContext(ContextUpdateType.All); // set initial context before loading any custom mod tokens
 
             // load context
             this.LoadContentPacks(contentPacks, installedMods);
-            this.UpdateContext(); // set initial context once patches + dynamic tokens + custom tokens are loaded
+            this.UpdateContext(ContextUpdateType.All); // set initial context once patches + dynamic tokens + custom tokens are loaded
 
             // register patcher
             helper.Content.AssetLoaders.Add(this.PatchManager);
@@ -256,7 +257,7 @@ namespace ContentPatcher
             helper.Events.Specialized.LoadStageChanged += this.OnLoadStageChanged;
 
             // set up commands
-            this.CommandHandler = new CommandHandler(this.TokenManager, this.PatchManager, this.PatchLoader, this.Monitor, this.RawContentPacks, modID => modID == null ? this.TokenManager : this.TokenManager.GetContextFor(modID), () => this.UpdateContext());
+            this.CommandHandler = new CommandHandler(this.TokenManager, this.PatchManager, this.PatchLoader, this.Monitor, this.RawContentPacks, modID => modID == null ? this.TokenManager : this.TokenManager.GetContextFor(modID), () => this.UpdateContext(ContextUpdateType.All));
             helper.ConsoleCommands.Add(this.CommandHandler.CommandName, $"Starts a Content Patcher command. Type '{this.CommandHandler.CommandName} help' for details.", (name, args) => this.CommandHandler.Handle(args));
 
             // can no longer queue tokens
@@ -277,15 +278,11 @@ namespace ContentPatcher
         }
 
         /// <summary>Update the current context.</summary>
-        /// <param name="affectedTokens">The specific tokens for which to update context, or <c>null</c> to affect all tokens</param>
-        private void UpdateContext(ConditionType[] affectedTokens = null)
+        /// <param name="updateType">The context update type.</param>
+        private void UpdateContext(ContextUpdateType updateType)
         {
-            InvariantHashSet onlyTokens = affectedTokens != null
-                ? new InvariantHashSet(affectedTokens.Select(p => p.ToString()))
-                : null;
-
-            this.TokenManager.UpdateContext(out InvariantHashSet changedGlobalTokens, onlyTokens);
-            this.PatchManager.UpdateContext(this.Helper.Content, changedGlobalTokens);
+            this.TokenManager.UpdateContext(out InvariantHashSet changedGlobalTokens);
+            this.PatchManager.UpdateContext(this.Helper.Content, changedGlobalTokens, updateType);
         }
 
         /// <summary>Load the registered content packs.</summary>
@@ -366,7 +363,8 @@ namespace ContentPatcher
                         foreach (KeyValuePair<string, ConfigField> pair in config)
                         {
                             ConfigField field = pair.Value;
-                            modContext.Add(new HigherLevelTokenWrapper(new ImmutableToken(pair.Key, field.Value, scope: current.Manifest.UniqueID, allowedValues: field.AllowValues, canHaveMultipleValues: field.AllowMultiple)));
+                            IValueProvider valueProvider = new ImmutableValueProvider(pair.Key, field.Value, allowedValues: field.AllowValues, canHaveMultipleValues: field.AllowMultiple);
+                            modContext.AddLocalToken(new Token(valueProvider, scope: current.Manifest.UniqueID));
                         }
 
                         // load dynamic tokens
@@ -431,7 +429,7 @@ namespace ContentPatcher
                                 values = new LiteralString("", localPath.With(nameof(entry.Value)));
 
                             // add token
-                            modContext.Add(new DynamicTokenValue(entry.Name, values, conditions));
+                            modContext.AddDynamicToken(entry.Name, values, conditions);
                         }
                     }